Final Push

+++ RMIClientSide/src/DeviceInterface.java
@@ -2,5 +2,5 @@
 import java.rmi.*;
 
 public interface DeviceInterface extends Remote{
-	public String addDevice(String name,double temp, int timestamp) throws RemoteException;
+	public String addDevice(String name,double temp, long timestamp) throws RemoteException;
 }
